otimização dos downloads

+++ baixa.py
@@ -11,13 +11,15 @@ from googleapiclient.errors import HttpError
 # If modifying these scopes, delete the file token.json.
 SCOPES = ['https://www.googleapis.com/auth/gmail.readonly']
 
-SECRETS_DIR = f'{os.path.abspath(os.path.dirname(__file__))}/secrets'
+ROOT_DIR = os.path.abspath(os.path.dirname(__file__))
+
+SECRETS_DIR = f'{ROOT_DIR}/secrets'
 TOKEN_JSON = f'{SECRETS_DIR}/token.json'
 CREDENTIALS_JSON = f'{SECRETS_DIR}/credentials.json'
-DOWNLOADS_DIR = f'{os.path.abspath(os.path.dirname(__file__))}/downloads'
+DOWNLOADS_DIR = f'{ROOT_DIR}/downloads'
 
 
-def main():
+def download(force=False):
     """Shows basic usage of the Gmail API.
     Lists the user's Gmail labels.
     """
@@ -56,13 +58,6 @@ def main():
 
                 for part in message['payload']['parts']:
                     if part['filename']:
-                        if 'data' in part['body']:
-                            data=part['body']['data']
-                        else:
-                            att_id = part['body']['attachmentId']
-                            att = messages_service.attachments().get(userId='me', messageId=message_id,id=att_id).execute()
-                            data = att['data']
-                        file_data = base64.urlsafe_b64decode(data.encode('UTF-8'))
                         file_name = part["filename"]
                         match = re.match(r'([0-9]{4})\_([0-9]{8})\.pdf', file_name)
                         if match:
@@ -70,16 +65,26 @@ def main():
                             file_year = file_date[4:8]
                             file_month = file_date[2:4]
                             file_day = file_date[:2]
-                            file_name = f'BOLETO-{file_year}-{file_month}-{file_day}'
+                            file_name = f'BOLETO-{file_year}-{file_month}-{file_day}.pdf'
                         else:
                             match = re.match(r'ESI\_([0-9]{4})\_([0-9]{2})\-([0-9]{4})\.pdf', file_name)
                             if match:
-                                file_name = f'EXTRATO-{match.group(3)}-{match.group(2)}'
+                                file_name = f'EXTRATO-{match.group(3)}-{match.group(2)}.pdf'
                         path = f'{DOWNLOADS_DIR}/{file_name}'
-                        print(f'Baixando {path}')
-                        with open(path, 'wb') as f:
-                            f.write(file_data)
-        
+                        if not os.path.isfile(path) or force:
+                            print(f'Baixando {file_name}')
+                            if 'data' in part['body']:
+                                data=part['body']['data']
+                            else:
+                                att_id = part['body']['attachmentId']
+                                att = messages_service.attachments().get(userId='me', messageId=message_id,id=att_id).execute()
+                                data = att['data']
+                            file_data = base64.urlsafe_b64decode(data.encode('UTF-8'))
+                            print(f'Escrevendo {file_name} em disco no caminho {path}')
+                            with open(path, 'wb') as f:
+                                f.write(file_data)
+                        else:
+                            print(f'Já existe {file_name}!')
 
     except HttpError as error:
         # TODO(developer) - Handle errors from gmail API.
@@ -87,4 +92,4 @@ def main():
 
 
 if __name__ == '__main__':
-    main()
+    download()
